Ensure column keys in virgin tables are unique on 32bit devices (#4051)
If column keys are not unique we end up finding the wrong backlink column during upgrade.

+++ src/realm/keys.hpp
@@ -103,7 +103,7 @@ struct ColKey {
         : value(val)
     {
     }
-    explicit ColKey(Idx index, ColumnType type, ColumnAttrMask attrs, unsigned tag) noexcept
+    explicit ColKey(Idx index, ColumnType type, ColumnAttrMask attrs, uint64_t tag) noexcept
         : ColKey((index.val & 0xFFFFUL) | ((type & 0x3FUL) << 16) | ((attrs.m_value & 0xFFUL) << 22) |
                  ((tag & 0xFFFFFFFFUL) << 30))
     {
